+++ src/components/Navigation/Navigation.test.tsx
@@ -7,7 +7,7 @@ import * as auth from "~/utils/auth"
 const useAuthSpy = vitest.spyOn(auth, "useAuth")
 
 import Navigation from "~/components/Navigation"
-import {mockUser} from "~/mocks/users"
+import {homer} from "~/mocks/users"
 
 test("shows unauthenticated routes", () => {
     useAuthSpy.mockReturnValueOnce(null)
@@ -19,8 +19,7 @@ test("shows unauthenticated routes", () => {
     )
 
     expect(screen.getByText("Home")).toBeInTheDocument()
-    expect(screen.getByText("About")).toBeInTheDocument()
-    expect(screen.getByText("Dashboard")).toBeInTheDocument()
+    expect(screen.queryByText("Notes")).not.toBeInTheDocument()
     expect(screen.getByText("Sign up")).toBeInTheDocument()
     expect(screen.getByText("Sign in")).toBeInTheDocument()
     expect(screen.queryByText("Sign out")).not.toBeInTheDocument()
@@ -28,7 +27,7 @@ test("shows unauthenticated routes", () => {
 })
 
 test("shows authenticated routes", () => {
-    useAuthSpy.mockReturnValueOnce(mockUser)
+    useAuthSpy.mockReturnValueOnce(homer)
 
     render(
         <MemoryRouter>
@@ -37,8 +36,7 @@ test("shows authenticated routes", () => {
     )
 
     expect(screen.getByText("Home")).toBeInTheDocument()
-    expect(screen.getByText("About")).toBeInTheDocument()
-    expect(screen.getByText("Dashboard")).toBeInTheDocument()
+    expect(screen.getByText("Notes")).toBeInTheDocument()
     expect(screen.queryByText("Sign up")).not.toBeInTheDocument()
     expect(screen.queryByText("Sign in")).not.toBeInTheDocument()
     expect(screen.getByText("Sign out")).toBeInTheDocument()