+++ src/mocks/users.ts
@@ -1,6 +1,6 @@
 import type {User} from "@prisma/client"
 
-const mockUser: User = {
+const homer: User = {
     id: 1,
     email: "homer@gmail.com",
     password: "password",
@@ -8,4 +8,14 @@ const mockUser: User = {
     updatedAt: new Date(),
 }
 
-export {mockUser}
+const marge: User = {
+    id: 2,
+    email: "marge@gmail.com",
+    password: "password",
+    createdAt: new Date(),
+    updatedAt: new Date(),
+}
+
+const mockUsers = [homer, marge]
+
+export {homer, marge, mockUsers}