+++ src/routes.ts
@@ -3,11 +3,14 @@ import {index, route} from "@remix-run/route-config"
 
 const routes: RouteConfig = [
     index("./routes/index.tsx"),
-    route("about", "./routes/about.tsx"),
-    route("dashboard", "./routes/dashboard.tsx"),
     route("signin", "./routes/signin.tsx"),
     route("signup", "./routes/signup.tsx"),
     route("signout", "./routes/signout.tsx"),
+    route("forgot", "./routes/forgot.tsx"),
+    route("reset/:token", "./routes/reset.tsx"),
+    route("notes", "./routes/notes.tsx"),
+    route("note/:id", "./routes/note.tsx"),
+    route("note/new", "./routes/newNote.tsx"),
     route("api/hello", "./routes/api/hello.tsx"),
     route("sentry", "./routes/sentry.tsx"),
     route("sentry/frontend", "./routes/sentryFrontend.tsx"),