+++ src/utils/auth.test.ts
@@ -2,17 +2,17 @@ import * as remix from "@remix-run/react"
 import {renderHook} from "@testing-library/react"
 import {describe, expect, test, vitest} from "vitest"
 
-import {mockUser} from "~/mocks/users"
+import {homer} from "~/mocks/users"
 import {useAuth} from "~/utils/auth"
 
 const useRouteLoaderDataSpy = vitest.spyOn(remix, "useRouteLoaderData")
 
 describe("useAuth", () => {
     test("uses authentication", () => {
-        useRouteLoaderDataSpy.mockReturnValueOnce({user: mockUser})
+        useRouteLoaderDataSpy.mockReturnValueOnce({user: homer})
 
         const {result} = renderHook(() => useAuth())
-        expect(result.current).toEqual(mockUser)
+        expect(result.current).toEqual(homer)
     })
 
     test("handles no user", () => {