+++ src/utils/session.server.test.ts
@@ -1,14 +1,14 @@
 import {createCookie} from "@remix-run/node"
 import {describe, expect, test, vitest} from "vitest"
 
-import {mockUser} from "~/mocks/users"
+import {homer} from "~/mocks/users"
+import * as users from "~/models/users.server"
 import {
     createSession,
     deleteSession,
     getSession,
     getUserFromSession,
 } from "~/utils/session.server"
-import * as users from "~/utils/users.server"
 
 const getUserByIdSpy = vitest.spyOn(users, "getUserById")
 
@@ -18,7 +18,7 @@ describe("createSession", () => {
 
         const response = await createSession({
             request,
-            userId: mockUser.id,
+            userId: homer.id,
             redirectUrl: "/",
         })
 
@@ -39,7 +39,7 @@ describe("createSession", () => {
 
         const response = await createSession({
             request,
-            userId: mockUser.id,
+            userId: homer.id,
             remember: true,
             redirectUrl: "/",
         })
@@ -61,31 +61,31 @@ describe("getSession", () => {
     test("gets session", async () => {
         const cookie = await createCookie("__session", {
             secrets: [process.env.SESSION_SECRET!],
-        }).serialize({userId: mockUser.id})
+        }).serialize({userId: homer.id})
 
         const request = new Request("https://example.com", {
             headers: {cookie},
         })
 
         const session = await getSession(request)
-        expect(session.get("userId")).toEqual(mockUser.id)
+        expect(session.get("userId")).toEqual(homer.id)
     })
 })
 
 describe("getUserFromSession", () => {
     test("gets user from session", async () => {
-        getUserByIdSpy.mockResolvedValueOnce(mockUser)
+        getUserByIdSpy.mockResolvedValueOnce(homer)
 
         const cookie = await createCookie("__session", {
             secrets: [process.env.SESSION_SECRET!],
-        }).serialize({userId: mockUser.id})
+        }).serialize({userId: homer.id})
 
         const request = new Request("https://example.com", {
             headers: {cookie},
         })
 
         const user = await getUserFromSession(request)
-        expect(user).toEqual(mockUser)
+        expect(user).toEqual(homer)
     })
 
     test("handles no user", async () => {
@@ -106,7 +106,7 @@ describe("deleteSession", () => {
     test("deletes session", async () => {
         const cookie = await createCookie("__session", {
             secrets: [process.env.SESSION_SECRET!],
-        }).serialize({userId: mockUser.id})
+        }).serialize({userId: homer.id})
 
         const request = new Request("https://example.com", {
             headers: {cookie},